# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Booking Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface Room {
  id: number;
  name: string;
  type: string;
  description: string;
  pricePerNight: number;
  maxGuests: number;
  imageUrl: string;
  amenities: string[];
  available: boolean;
  rating: number;
  reviewCount: number;
}

export interface CreateRoomRequest {
  name: string;
  type: string;
  description: string;
  pricePerNight: number;
  maxGuests: number;
  imageUrl: string;
  amenities: string[];
  available: boolean;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  confirmed: "confirmed",
  cancelled: "cancelled",
  pending: "pending",
} as const;

export interface Booking {
  id: number;
  userId: number;
  roomId: number;
  checkIn: string;
  checkOut: string;
  guests: number;
  totalPrice: number;
  status: BookingStatus;
  createdAt: string;
  room?: Room;
}

export type BookingWithDetailsStatus =
  (typeof BookingWithDetailsStatus)[keyof typeof BookingWithDetailsStatus];

export const BookingWithDetailsStatus = {
  confirmed: "confirmed",
  cancelled: "cancelled",
  pending: "pending",
} as const;

export interface BookingWithDetails {
  id: number;
  userId: number;
  roomId: number;
  checkIn: string;
  checkOut: string;
  guests: number;
  totalPrice: number;
  status: BookingWithDetailsStatus;
  createdAt: string;
  room?: Room;
  user?: User;
}

export interface CreateBookingRequest {
  roomId: number;
  checkIn: string;
  checkOut: string;
  guests: number;
}

export type GetRoomsParams = {
  checkIn?: string;
  checkOut?: string;
  guests?: string;
  roomType?: string;
};