# lib/api-zod/src/generated/types/bookingWithDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Booking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingWithDetailsStatus } from "./bookingWithDetailsStatus";
import type { Room } from "./room";
import type { User } from "./user";

export interface BookingWithDetails {
  id: number;
  userId: number;
  roomId: number;
  checkIn: string;
  checkOut: string;
  guests: number;
  totalPrice: number;
  status: BookingWithDetailsStatus;
  createdAt: string;
  room?: Room;
  user?: User;
}