# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Booking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingRequest {
  roomId: number;
  checkIn: string;
  checkOut: string;
  guests: number;
}