# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hotel Booking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: number;
  name: string;
  type: string;
  description: string;
  pricePerNight: number;
  maxGuests: number;
  imageUrl: string;
  amenities: string[];
  available: boolean;
  rating: number;
  reviewCount: number;
}